changed some database schemas - update script for object-table

+++ webEdition/liveUpdate/sqldumps/tblSchedule.sql
@@ -1,5 +1,5 @@
 CREATE TABLE ###TBLPREFIX###tblSchedule (
-  DID bigint unsigned NOT NULL default '0',
+  DID int unsigned NOT NULL default '0',
   `expire` DATETIME NOT NULL,
   `lockedUntil` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	`task` enum('publish','park','delete','doctype','category','directory','search_enable','search_disable','call') default 'publish',
