+++ webEdition/liveUpdate/sqldumps/tblcustomerfilter.sql
@@ -10,7 +10,7 @@
 /* query separator */
 
 CREATE TABLE ###TBLPREFIX###tblcustomerfilter (
-  modelId bigint unsigned NOT NULL,
+  modelId int unsigned NOT NULL,
   modelType enum('objectFile','image/*','text/html','text/webedition','folder','application/x-shockwave-flash','application/*','video/*','audio/*')  NOT NULL default 'application/*',
   modelTable enum('tblFile','tblObjectFiles') NOT NULL,
   accessControlOnTemplate tinyint unsigned NOT NULL default '0',
