+++ webEdition/liveUpdate/sqldumps/tblformmailblock.sql
@@ -1,5 +1,5 @@
 CREATE TABLE ###TBLPREFIX###tblformmailblock (
-  id bigint unsigned NOT NULL auto_increment,
+  id mediumint unsigned NOT NULL auto_increment,
   ip varchar(40) NOT NULL,
   blockedUntil int NOT NULL,
   PRIMARY KEY  (id),
