+++ webEdition/liveUpdate/sqldumps/tblversions.sql
@@ -55,7 +55,7 @@ CREATE TABLE ###TBLPREFIX###tblversions (
   Templates tinytext NOT NULL,
   MasterTemplateID int unsigned NOT NULL default '0',
   TableID mediumint unsigned NOT NULL,
-  ObjectID bigint unsigned NOT NULL,
+  ObjectID int unsigned NOT NULL,
   IsClassFolder tinyint unsigned NOT NULL,
   Charset ENUM('','UTF-8','ISO-8859-1','ISO-8859-2','ISO-8859-3','ISO-8859-4','ISO-8859-5','ISO-8859-6','ISO-8859-7','ISO-8859-8','ISO-8859-9','ISO-8859-10','ISO-8859-11','ISO-8859-12','ISO-8859-13','ISO-8859-14','ISO-8859-15','Windows-1251','Windows-1252') NOT NULL default '',
   active tinyint unsigned NOT NULL,
